chore: use APPDATA for .terramate.d directory on windows
